Rename law submodule as spectral_law
