+++ skrmt/ensemble/__init__.py
@@ -8,10 +8,10 @@ from .wishart_ensemble import WishartEnsemble
 from .manova_ensemble import ManovaEnsemble
 from .circular_ensemble import CircularEnsemble
 
-from .law import WignerSemicircleDistribution
-from .law import MarchenkoPasturDistribution
-from .law import TracyWidomDistribution
-from .law import ManovaSpectrumDistribution
+from .spectral_law import WignerSemicircleDistribution
+from .spectral_law import MarchenkoPasturDistribution
+from .spectral_law import TracyWidomDistribution
+from .spectral_law import ManovaSpectrumDistribution
 
 from .plot_law import wigner_semicircle
 from .plot_law import marchenko_pastur
