+++ skrmt/ensemble/plot_law.py
@@ -14,7 +14,7 @@ import matplotlib.pyplot as plt
 from .gaussian_ensemble import GaussianEnsemble
 from .wishart_ensemble import WishartEnsemble
 from .manova_ensemble import ManovaEnsemble
-from .law import (
+from .spectral_law import (
     WignerSemicircleDistribution,
     MarchenkoPasturDistribution,
     TracyWidomDistribution,
