+++ skrmt/ensemble/spectral_law.py
